requirements.txt
streamlit==1.19.0
pandas==1.5.3
matplotlib==3.5.1
numpy==1.21.5
scikit-learn==0.24.2
umap-learn==0.5.3
plotly==5.13.0
graphviz==0.20.1

